+++ test/util.py
@@ -7,3 +7,12 @@ __config = LocalConf(DEFAULT_CONFIG)
 # Base config to use when mocking
 def base_config():
     return deepcopy(__config)
+
+
+class Anything:
+    """Class matching any object.
+
+    Useful for assert_called_with arguments.
+    """
+    def __eq__(self, other):
+        return True
